move `getFunctionSelector` into `functionSignatureUtils`
